collection of fixes and changes to WC

+++ app/src/main/java/com/concordium/wallet/data/walletconnect/Payload.kt
@@ -5,8 +5,8 @@ sealed interface Payload {
     data class ContractUpdateTransaction(
         val address: ContractAddress,
         val amount: String,
-        var maxEnergy: Int,
-        var maxContractExecutionEnergy: Int,
+        var maxEnergy: Long,
+        var maxContractExecutionEnergy: Long,
         val message: String,
         val receiveName: String
     ) : Payload
